Inline the view-config helper into the Navigation list route
The shared routes/view-config workspace had a single consumer. Fold it into
routes/navigation-list/view-utils.ts until a second screen needs it.
