+++ routes/navigation-list/package.json
@@ -21,7 +21,6 @@
 		"@wordpress/notices": "file:../../packages/notices",
 		"@wordpress/route": "file:../../packages/route",
 		"@wordpress/routes-lock-unlock": "file:../lock-unlock",
-		"@wordpress/routes-view-config": "file:../view-config",
 		"@wordpress/views": "file:../../packages/views"
 	}
 }
