--- routes/view-config/package.json
@@ -1,15 +1,0 @@
-{
-	"name": "@wordpress/routes-view-config",
-	"version": "1.0.0",
-	"private": true,
-	"exports": {
-		".": "./index.ts",
-		"./package.json": "./package.json"
-	},
-	"dependencies": {
-		"@wordpress/core-data": "file:../../packages/core-data",
-		"@wordpress/data": "file:../../packages/data",
-		"@wordpress/dataviews": "file:../../packages/dataviews",
-		"@wordpress/routes-lock-unlock": "file:../lock-unlock"
-	}
-}
