--- routes/view-config/tsconfig.json
@@ -1,19 +1,0 @@
-{
-	"$schema": "https://json.schemastore.org/tsconfig.json",
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"jsx": "react-jsx",
-		"rootDir": ".",
-		"noEmit": true,
-		"emitDeclarationOnly": false,
-		"composite": false,
-		"types": []
-	},
-	"references": [
-		{ "path": "../../packages/core-data/tsconfig.build.json" },
-		{ "path": "../../packages/data" },
-		{ "path": "../../packages/dataviews/tsconfig.build.json" }
-	],
-	"include": [ "**/*.ts", "**/*.tsx" ],
-	"exclude": [ "build", "node_modules" ]
-}
